requirements.txt
# Omega Super Desktop Console - Initial Prototype Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
pyjwt>=2.8.0
cryptography>=41.0.0
psutil>=5.9.0
asyncio-mqtt>=0.16.0
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
numpy>=1.25.0
scikit-learn>=1.3.0
torch>=2.1.0
onnx>=1.15.0
onnxruntime>=1.16.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
protobuf>=4.25.0
websockets>=12.0
aiofiles>=23.2.1
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0
typer>=0.9.0
